Move caldav coordinator to its own file.
